Handle pagination of results
Previously, the code was only checking the first page of results from
the API, which is 100 images.  This provides a lot of opportunity for
incorrect results when the number of images in a repository grows larger
than 100.

Now we create a paginator object and use that to iterate through all of
the images.  This may take longer, especially for non-existent tags,
since we have to make more API calls, but it's necessary for accuracy.

Relevant documentation:
  * https://aws.github.io/aws-sdk-go-v2/docs/making-requests/#using-paginators
  * https://pkg.go.dev/github.com/aws/aws-sdk-go-v2/service/ecr#ListImagesPaginator